OpenAI Image Bot/requirements.txt
discord
python-dotenv
openai
pychatgpt == 0.4.3.3